refactor: modifying driver schema and changing model -> schema
