+++ backend/controllers/adminController.js
@@ -1,7 +1,12 @@
 const asyncHandler = require("express-async-handler");
 const bcrypt = require("bcryptjs");
-const Admin = require("../models/adminModel");
 const jwt = require("jsonwebtoken");
+const Admin = require("../models/adminSchema");
+
+// Generate Token
+const generateToken = (id) => jwt.sign({ id }, process.env.JWT_SECRET, {
+    expiresIn: "30d",
+  });
 
 // @desc Get all admins
 // @route /api/admin
@@ -141,7 +146,7 @@ const deleteAdmin = asyncHandler(async (req, res) => {
 const loginAdmin = asyncHandler(async (req, res) => {
   const { email, password } = req.body;
 
-  //get email
+  // get email
   const foundAdmin = await Admin.findOne({ email });
 
   if (foundAdmin && (await bcrypt.compare(password, foundAdmin.password))) {
@@ -170,13 +175,6 @@ const getAdmin = asyncHandler(async (req, res) => {
   };
   res.status(200).json(admin);
 });
-
-// Generate Token
-const generateToken = (id) => {
-  return jwt.sign({ id }, process.env.JWT_SECRET, {
-    expiresIn: "30d",
-  });
-};
 
 module.exports = {
   getAdmins,
