+++ backend/controllers/logsController.js
@@ -1,7 +1,7 @@
 const PickupLog = require("../models/pickupLogSchema");
 const DropoffLog = require("../models/dropoffLogSchema");
-const Admin = require("../models/adminModel.js");
-const Driver = require('../models/driverModel.js')
+const Admin = require("../models/adminSchema");
+const Driver = require("../models/driverSchema");
 
 // @desc Get all pickup logs
 // @route GET : /api/pickup
@@ -180,7 +180,7 @@ const pushPickups = async (req, res) => {
 
   // Get array from request body
   const data = req.body;
-  let counter = 0;
+
   try {
     const response = await PickupLog.insertMany(data, { ordered: false });
      res.status(200).json(response);
