+++ backend/models/vehiclesSchema.js
@@ -1,15 +1,21 @@
-const mongoose = require('mongoose')
+const mongoose = require("mongoose");
 
-const VehicleSchema = new mongoose.Schema({
+const VehicleSchema = new mongoose.Schema(
+  {
     name: String,
     driver: String,
     isLoggedIn: Boolean,
     img: String,
     currentPickups: Array,
     currentDropoffs: Array,
-    totalWeight: Number,
-}, {collection: 'Vehicle'});
+    totalFoodAllocation: {
+      type: Map,
+      of: Number,
+    },
+  },
+  { collection: "Vehicle" }
+);
 
-const Vehicle = mongoose.model('Vehicle', VehicleSchema)
+const Vehicle = mongoose.model("Vehicle", VehicleSchema);
 
-module.exports = Vehicle
+module.exports = Vehicle;
